vcf-etl/test_from_yumosong/pdm_test/pyproject.toml
[project]
name = "pdm_test"
version = "0.1.0"
description = ""
authors = [
    {name = "yumosong", email = "songyumo5@gmail.com"},
]
dependencies = [
    "PyVCF>=0.6.8",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
